requirements.txt
ultralytics
opencv-python
pyyaml
numpy
